Bug#30453221: CHECK CONSTRAINTS CHARACTER SET MISMATCH
CREATE TABLE with generated column expressions (the CHECK constraint
expression in the bug report) referencing a non-ASCII identifier
caused a syntax error if the current client character set was not compatible
with UTF-8 (e.g. GBK).

Analysis
========

While executing the CREATE TABLE statement containing expressions like
the CHECK constraint,
MySQL reconstructs the source string of that constraint's expression
from its AST. Then MySQL re-parses that reconstructed expression source
immediately. While the reconstructed expression string is encoded with
utf8mb3, it was re-parsed using the current client character set.
If the reconstructed expression string contained byte sequences not
valid in the current client character set, that caused failures.

Solution
========

Substitute the current client character set with the system charset (utf8mb3)
while re-parsing reconstructed source string of generated column expressions.

This had a fall-out: a test in functional_index.test failed:

   CREATE TABLE t1(f1 JSON, INDEX idx1 ((CAST(f1->>"$.name" AS CHAR(30)) COLLATE utf8mb4_bin)));
   :
   EXPLAIN SELECT * FROM t1 WHERE CAST(f1->>"$.name" AS CHAR(30)) COLLATE utf8mb4_bin = "james";

no longer picked the index. This was due to the fact that when
comparing the generated column defintion with the left side of the
equality expression in the SELECT, we eventually compare the arguments
of COLLATE, i.e.  utf8mb4_bin vs. utf8mb4_bin in get_gc_for_expr. They
are the same, and should match, but do not. The reason is that the
generated column text ('utf8mb4_bin') is now parsed with system with
and gets the collation of character set (utf8mb3) whereas the
corresponding string (also utf8mb4_bin') in the query gets resolved to
the current session character set (utf8mb4).

To avoid this issue, we give Item_func_set_collation its own Item::eq
implementation in lieu of its old Item::eq_specific, which only checks
that the resolved collations are the same, i.e. not worrying about the
character set of the collation strings themselves.  It's not really a
string seen from the user's point of view; although it can be given as
a string, it can not be given a char set prefix. This fixed the issue
in functional_index.

Note that even before the present patch, this weakness could rear its
ugly head, say be doing:

   CREATE TABLE t1(f1 JSON, INDEX idx1 ((CAST(f1->>"$.name" AS CHAR(30)) COLLATE utf8mb4_bin)));
   SET charset utf8mb3;
   EXPLAIN SELECT * FROM t1 WHERE CAST(f1->>"$.name" AS CHAR(30)) COLLATE utf8mb4_bin = "james";

would not pick the index either. This now works as expected as well.

Change-Id: I9786e0b458975a5f6d5e5cb14d2fe49d70a46506