Bug#37773000	Debug build router crashes when executing mysql_task MTR test
In Debug server builds, calling mysql_real_query_nonblocking() on a connection
in error state (ie. last mysql_errno() returned a client error) aborts on an
assert check.

Router task management code was changed to not try to perform cleanup if the
last query failed with an error in range 2XXX.

Change-Id: I8a3adfb922b913ac79d5d30cbf412d9f0e2859d3