Bug#36784795 Query rewrite plugin not working when server runs with                           autocommit=OFF
Query rewrite plugin reload rules logic runs in an independent thread
created when a request is submitted. This thread runs the reload logic
and exits. When the server is in global autocommit OFF mode, the new
thread and its session will also run with autocommit OFF mode. At the
end of the reload execution, the code does a statement commit and exist.
Statement commits does not preserve the changes in autocommit OFF
mode. Due to this, after the thread exits, any table data changes done
will be lost.

Fix: Perform a transaction commit to ensure all changes are preserved
before exiting the session/thread.

Change-Id: Ib6f96873f90d99b2c53bb0157e91b160760b933e