Bug#37788926 Router crash if session is logged out while async task running at router
Async tasks executed at the Router (as opposed to at the Server
using mysql_tasks) would crash if the session that started the
task logs out while the task is running.
This is because the session pool is destroyed at log out, but
when the task finishes it is returned to the now invalid pool.
Async tasks executed at the Router now hold a shared reference
to the pool until the task is done.

Change-Id: I11a03be7873e8f1f92075af55f333ccc82a8e9f7