Bug#37436220 MySQL Server SEGV within JOIN::refresh_base_slice
In the simplified repro, we have a column which is defined
with two default value expressions:

   DEFAULT ((SELECT 'string')) DEFAULT (FALSE)

where the second one overrides the first. However, the subquery is
added to the QE/QB structure and is not removed when we override the
default value with FALSE. If there were no second default, we would
normally see an error message, since a subquery is not allowed in the
default expression:

   ER_DEFAULT_VAL_GENERATED_FUNCTION_IS_NOT_ALLOWED

However, the checking is only performed after the second default
overrides the first, so no checking is never performed on the first
default, even though is is contextualized.  If the create statement
also has AS SELECT, we see an assert in debug builds because the
subquery is attemped optimized, but has never been prepared (SIGSEGV
in production builds).

Solution: eliminate overridden column attributes before
contextualizing. Which attributes override textually preceding
occurences and which are additive (e.g. like CHECK constraints) should
be better documented.

Change-Id: I1730b180ba9de55c30f2f27b7dd850a47ed4b9f6