Bug#35521787 nth_value window function call assertion error [part 2 of 2]
The bug report has a second repro which fails at the same assert, but
is due to another error fixed here.

RANGE unit comparison failed due to underflow when computing
after_frame/before_frame: we had an unsigned column with a this frame
specification:
  :
  ORDER BY t0.c1 DESC
  RANGE BETWEEN CURRENT ROW AND 10 FOLLOWING)

the values in the frame, when sorted descendingly were
  371    first row
  2      second row

when computing the frame limit, we have

     BETWEEN CURRENT ROW AND 10 FOLLOWING
i.e. BETWEEN 2 AND 2 - 10
i.e. BETWEEN 2 AND -8

which be an underflow for an unsigned int. This was not detected, and
bad things happened. This is a weakness of the current implementation
at the range borders.  We now give an error message instead of the
assert and/or wrong result.

Change-Id: I1c794a0dbb25cd399c9525782b9117cfe0082be7