Bug#37791945: Remove NODISCARD macro
* Background

The NODISCARD macro, defined in the header nodiscard.h, was introduced
to work around gcc bug
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=84476. In this bug, gcc
does not emit warnings as expected for some virtual function calls,
despite using the built-in [[nodiscard]] attribute. Therefore, on gcc,
the macro expands to gnu::warn_unused_result which is a gcc-specific
function attribute that has almost the same meaning, but does not have
the problem with virtual function calls. On other compilers the macro
expands to nodiscard.

* Problem

It has turned out that gnu::warn_unused_result has other bugs or
problems.

On functions declared to return auto, where the actual return type
resolves to void, gnu::warn_unused_result generates a compilation
warning (which may get promoted to error) because of having the
attribute on a void function, whereas nodiscard allows this. That
makes the attribute unusable on such functions, as there is no way for
the caller to use a void return value. Therefore, nodiscard is better
in such generic code.

Also, gnu::warn_unused_result appears to impact how gcc analyzes use
of uninitialized memory. We could not define the exact problem, but
using nodiscard instead of gnu::warn_unused_result removed what
appears to be a wrong compilation warning.

* Fix

Use [[nodiscard]] instead of [[NODISCARD]] and remove nodiscard.h.

Despite the bug in nodisard, we can't forget checking return values,
beacuse we build with clang too.

Change-Id: I017fd2ea4f6e0460fc3d2fdc09ece708b8a6bbd8