WL#16076 step 16: utils library: new call_and_catch feature
WL#16076: New GTID set data structure

Introduce the the header call_and_catch.h in the utils library. This
contains a wrapper function that calls a given invocable, catches
exceptions from it, and returns a value that indicates if an exception
was caught. If the invokable returns void, call_and_catch returns
`Return_status::error`; otherwise it returns `std::optional<T>` where
`T` is the return type from the invokable (possibly wrapped in a
reference_wrapper, if necessary).

Change-Id: I5e9c9200e5b9cbe69018c720af7dc366d23af973