WL#16790: Refactor TIME handling in server
This patch changes internal handling of TIME values in the MySQL server
from using the MYSQL_TIME struct to use the new class Time_val.

Class Time_val is designed to be small and efficient, and to improve
maintainability of temporal types in the server.

See Low Level Design of the worklog for details.

Change-Id: I2ae6cc79cf25d6a4ce995e1b876049e103beb78e