Bug#36182490 MIN, as window function, returns wrong NULL result
Wrong logic in Item_sum_hybrid::compute for case
We have four cases:
               m_want_first  m_nulls_first
          (1)       F            F
          (2)       F            T
          (3)       T            F
        * (4)       T            T

In this case, we have a non-buffered MIN functions (so optimized
because the frame is ordered ascendingly on the aggregate's argument
'a' field which is also nullable): ORDER BY a ROWS UNBOUNDED PRECEDING
Rows:
     NULL   -> MIN(a) should be NULL
     1      -> MIN(a) should be 1
     2      -> MIN(a) should be 1
     :

In row one, there is a NULL, so the MIN window function correctly
returns NULL, too. However, for row 2, we have the non-null value 1.
That is, the frame contains both a NULL and 1. A bug in the logic
leads to this not being detected, so we get MIN(arg) == NULL for all
succeeding rows.

Solution: fix logic
Change-Id: I3591eb7ccc2c225f35de09e7590968ffbede8c42