Bug#37733272: Planning overhead for ordered range queries in hypergraph [6, skip scan subsumed predicates, noclose]
This patch does not reduce the planning overhead, but it improves the
quality of the index skip scan plans considered by the hypergraph
optimizer. This allows a later patch to reduce the number of index
skip scans to consider, and thereby reducing the planning time,
without sacrificing plan quality.

Currently, the hypergraph optimizer is not able to avoid redundant
filters after index skip scans that use more than one predicate. To
compensate for that, it tries to generate index skip scans for every
single predicate in the WHERE clause, in the hope that it finds a
single-predicate skip scan that is able to subsume the predicate and
avoid the filter. This is time-consuming, and most of the extra plans
it considers are not very good.

For example, with an index on (a,b,c) and the query

  SELECT 1 FROM t WHERE a = 1 AND c = 1

it will propose an index skip scan based on the full WHERE clause with
a redundant filter on top. Additionally, it will propose a skip scan
for only the c = 1 predicate, with a filter for a = 1 on top. The
latter plan does not have a redundant filter, but it is usually much
slower because it has to step through all values of the a column in
the index, and that cost typically overshadows the cost of the
redundant filter.

It would be better that the skip scan plan for the full WHERE clause
avoided the redundant filter, so that the optimizer didn't have to
spend time proposing inferior skip scans to avoid the filter.

This patch adds logic for analyzing which predicates are subsumed by
an index skip scan. It does this by reusing the function
FindAbsorbedPredicatesForRangeScan(), which performs the same analysis
for regular index range scans.

Change-Id: I857ce7f766d636f510aeeb45336442b4778bdc6e