Bug#37810512 JSON results are encoded as strings in async task function/procedure endpoints
When function/procedure endpoints that return JSON or VECTOR data were
called, results would be returned either as strings instead of inlined
JSON or error out, in case of VECTORs.

The cause is two-fold:
- @variables only support a few basic types, so if a JSON value is
assigned to a @var, it will be implicitly cast to string.
- JSON_OBJECT() does not support vector data (unlike virtually every
other type), resulting in an error if a vector value is passed to it.

This patch:
- explicitly casts/converts @ OUT params to JSON when building the JSON
response object for procedures
- explicitly casts/convers function return values to JSON when building
the response for functions

Change-Id: If7e7312a853c32dadc7617b5c1f6d68c56d0e10a