Bug#37733272: Planning overhead for ordered range queries in hypergraph [5, refactor applied predicates, noclose]
This is a refactoring to simplify patches that come later.

The code that finds the fields of the used key parts in an index range
scan is factored out of FindAppliedAndSubsumedPredicatesForRangeScan()
and moved into a separate function called
FindAbsorbedFieldsForRangeScan(). This is done in order to make it
easier to reuse the former function for analyzing the predicates in an
index skip scan. FindAppliedAndSubsumedPredicatesForRangeScan() is
also renamed to FindAbsorbedPredicatesForRangeScan() for brevity, and
it is made to return a class called AbsorbedPredicates which contains
the two bitsets, instead of returning them in two separate output
parameters, to make it easier to pass the two bitsets around.

Change-Id: If20c5b9c6b2c3b12314f858cf5be0572971d04c1