Bug#37733272: Planning overhead for ordered range queries in hypergraph [7, avoid bad skip scans, noclose]
The hypergraph optimizer spends unnecessary time on proposing index
skip scan plans that are unlikely to be any better than other plans.

For example, if there is an index on (a,b,c), a query such as

  SELECT 1 FROM t WHERE a = 1 AND b = 1 AND c = 1

will propose one index skip scan using the predicate b = 1, and
another one using the predicate c = 1. Both of these plans are clearly
inferior to an index lookup on (a,b,c)=(1,1,1).

The only reason to propose these plans, was that the hypergraph
optimizer was not capable to remove redundant predicates on the index
skip scans that used multiple predicates. This was fixed in the
previous patch, so there no longer any benefit in proposing them.

This patch removes the code in the hypergraph optimizer that collects
index skip scans for each single predicate. It now collects index skip
scans only based on the full WHERE clause. This is the same as what
the old optimizer does.

The change removes a good chunk of the planning time for queries such
as the one in the example above.

Change-Id: I1cd1741e35a5a3f2b1a94df320fa10d4287e1811