Bug#37563088: mysql_secure_installation does not check for expired passwords anymore
mysql_secure_installation was co-operating with the installer to deal
with expired passwords.
The installer removed this support some time ago. And (part of) the
m_s_i support was removed too by bug #28235716. But some of it was left
behind, resulting in dead code.
Removed the rest of the code now.

Change-Id: I2fac25a62c86b8375ce2e2f332c84ff08e802193