Bug#37833212 Limit the number of CPUs involved for cmake targets 'run-profile-suite' and 'test-unit'
The default parallelism for running 'run-profile-suite' and
'test-unit' is the number of CPUs of the current host. This does not
work well for machines with hundreds of CPUs.  Use the predefined
environment variables MTR_MAX_PARALLEL and CTEST_PARALLEL_LEVEL to
limit parallelism for mtr and ctest respectively.

MTR_MAX_PARALLEL and CTEST_PARALLEL_LEVEL can also be set on the cmake
command line.

Change-Id: I69041949210fb5b165dd5b9d2d687eda6460cf04