Bug#37824096 - passthroughDbUser works with cookies but not JWT
Description
===========
When a MySQL Internal user logs in using cookies with passthroughDbUser enabled,
a dedicated MySQL connection pool is created for that user and is used
when serving regular requests for that session.
However, when the session type is JWT, the MySQL connection pool ends up empty.

Fix
===
While authenticating, MRS session is dropped at the end of authentication when
user tries to generate JWT. New session is created to access user data when
JWT is used against REST resources.
Reuse MRS session that is used for authenticating the user.

Change-Id: I82f015534bd7ccebd7f8f00ebc589e79b3f12680