Bug#37480474  Debug ut_error in  dd_is_table_in_encrypted_tablespace for discarded tablespace
Symptom:
Sometimes an operation which opens a table for which the tablespace has
been discarded will terminate the server in debug builds.

Cause:
When a table is opened, ha_innobase::open calls
dd_is_table_in_encrypted_tablespace, which attempts to obtain information
about the tablespace. When the tablespace has been discarded, it is not
present in the filesystem, and subsequently the relevant entry is removed
from InnoDB's cache. In this case the lookup falls back on the data
dictionary, which normally holds information about discarded tablespaces.
However if the table is opened while an action is being performed which
triggers a rebuild, even though there is no actual tablespace to rebuild,
the process still updates the data dictionary by removing the old entry
for the tablespace and inserting a new one. This operation is performed in
stages, where the old tablespace is first marked for deletion, then MDL
is downgraded to allow concurrent access to the table while it's being
rebuilt. When this happens, attempts to find the DD row for the tablespace
using READ_UNCOMMITTED isolation level (because that's what the function
Dictionary_client::acquire_uncached_uncommitted_impl<dd:Tablespace>, which
is called for this lookup, uses) will fail to find the row for the
tablespace.

Fix:
Check if the tablespace is discarded and short-circuit before attempting to
retrieve the tablespace object. The check for tablespace encryption is not
needed in this case anyway as the whole condition requires the tablespace
not to be discarded.

Change-Id: Id0ee0b311d07596a5bfd1a6385d7efd40b68b656