Bug#37822992 Duplicates may stop the server when rebuilding PK
Description:
------------
Possiblity of stopping the server when duplicate error
is reported when primary key is rebuilt.

Analysis:
---------
When duplicate is identified during index rebuild previous
row is being used to report it. The problem is that due to
some optimizations only key values are there. This is not
enough for report logic. In addition, finish callback is
called after error happened during index rebuild. It
attempts to finalize the process after load tasks were already
released and finished during error handling.

Fix:
----
Using current row to report duplicate, because it has all fields.
Do not attempt to call finish callback after error was reported.

This patch is based on contribution from Xizhe Zhang at Alibaba Inc.

Change-Id: Ib29c2f608e9a52fb17adef01795d9b8b3fb4ebfe