Bug#37851010 MRS Async Procedure/Function REST Endpoints miss handling of session variables
Added a new mysqlTask.userVariables array option for routine endpoints.
The option takes a list of @user_variable names that cause matching
object_field parameter descriptors for routines to be interpreted as
user variables instead of regular parameters. "in" and "inout" values
are takesn from the input payload from the request and "out" and "inout"
values are included in the response object.

Change-Id: If1044a4a0b6feb845275adc760fb2f55a8a1df4c