Bug#37808260: Performance degradation in MySQL>=8.4 vs MySQL 8.0               due to redundant condition not optimized away
This is regression from the fix for Bug#30112096. As part of the fix,
Item_row::compile was introduced. This made constant propagation
apply for "IN" predicates having more than one column on the
left hand side. For the query in the bugpage, the equality for one
of the column is propagated to the IN predicate. This makes it not
eligible for range analysis. Range optimizer expects all expressions
on the left hand side of the IN predicate to be columns from a table.
Since, it is now transformed into a constant, it rejects this predicate.
As a result, the impossible where condition is not detected resulting
in a less performant query.

Fix is to disable constant propagation for IN predicates having more than
one column on its LHS.

Change-Id: I91a6fd0f4c52fc64e463a7d94d3555821bfbfd6b