Bug#37733272: Planning overhead for ordered range queries in hypergraph [8, pretty print, noclose]
Refactoring to produce easier to understand error messages from unit
test failures in hypergraph_optimizer-t. It factors out into a helper
function code that pretty prints AccessPath type names in the
optimizer trace, so that it can be reused in the unit test.

Example test failure before the patch:

  Expected equality of these values:
    AccessPath::INDEX_SCAN
      Which is: 
    root->type
      Which is: \0

The same test failure after the patch:

  Expected equality of these values:
    AccessPath::INDEX_SCAN
      Which is: INDEX_SCAN
    root->type
      Which is: TABLE_SCAN

Change-Id: Id8544fcb94f12a8b5bbd256d6a925b8a0b04b678