Bug#37733272: Planning overhead for ordered range queries in hypergraph [9, index scan, noclose]
When table access can be performed using a non-parameterized index
lookup, the hypergraph optimizer also considers a full index scan as
an alternative. The full index scan is never cheaper than the index
lookup, as it has to read at least as many rows from the table, and
often it is much more expensive.

When a non-parameterized index lookup is found, it is therefore a
waste of time to explore the full index scan alternative. This patch
makes the hypergraph optimizer propose full index scans only if it
hasn't found a non-parameterized index lookup on the same index.

Change-Id: Ib64dff2e732b5d2dbfa1bfeddc85930454d3f6c0