Bug#37371443: Crash at at /opt/rh/gcc-toolset-12/root/usr/include/c++/12/bits/new_allocator.h:158
Fix SIGSEGV in vector due to internal memory management issue

The vector's reallocation process was causing a SIGSEGV due to a deallocation
issue. This occurred when the vector resized itself to accommodate more
elements. The issue is likely related to the vector's internal memory
management.

Resolution: Pre-allocate memory using reserve() to prevent unnecessary
reallocations and likely fix the SIGSEGV.

Change-Id: Ic80031e161aed5d56289fd0df484d93a08db839a