Bug#37733272: Planning overhead for ordered range queries in hypergraph [10, table scan, noclose]
The hypergraph optimizer spends time exploring alternatives using
table scans even though it has found plans that are known to be as
fast or faster than a table scan.

When a covering non-parameterized index lookup or a covering index
scan has been found for a table, a table scan is not a cheaper
alternative. The table scan will have to read at least the same number
of rows, and it will have to read at least the same amount of data.

This patch makes the hypergraph optimizer stop considering table scans
if it has found that it can use a covering index instead.

Change-Id: Ia999c075c5245bbc6712b5b0f3b7062b89f542d5