Bug#37733272: Planning overhead for ordered range queries in hypergraph [11, final predicates, noclose]
The hypergraph optimizer has logic to skip the step that applies final
predicates to the plan if there are no final predicates. (The final
predicates are the predicates that are either non-deterministic or
don't refer to any tables). It is currently only skipped for point
select queries, but it takes a noticeable chunk of time even for other
simple queries.

This patch makes the hypergraph optimizer skip this step for all
queries that don't have any final predicate, to help speed up simple
single-table queries, even if they use a non-unique index lookup
instead of the unique index lookup used in point selects.

Change-Id: I6c5157c8ed952ed0593dcc556dd0f7eb14b7bc16