Bug#37859830: rapid.pfs_secondary result mismatch with --hypergraph flag
When the hypergraph optimizer is enabled for both the primary engine
and the secondary engine, a statement that is prepared for the primary
engine is not always automatically reprepared if a subsequent
execution is offloaded to the secondary engine.

When using the old optimizer in the primary engine and the hypergraph
optimizer in the secondary engine, repreparation happened
automatically when changing engines, but only because there was code
to detect that a different optimizer was used. When the same optimizer
is used by both engines, it wasn't always detected that a
repreparation was needed.

Fixed by extending the check during execution to also check which
engine the statement was prepared for, and repreparing if a different
engine is used for execution.

Change-Id: I96b478e16b656914ff2ec31eb182fa9bbee18a8b