Bug#37772498 Refactor contention mode in JitExecutor Context Pool [post-fix]
The heap monitoring APIs used to determine the context creation
capacity is not available on Windows, thus caused a denial of contexts
for JavaScript processing. The problem is resolved by using a less
flexible contention mode in windows, creation of contexts will be
available all the time until it fails once, in which case, the contention
mode will be turned on and only reuse of the existing contexts will be
possible.

As an improvement, a new struct was defined to hold the values that can
be used to configure the memory limits in the jit-executor. Formatting
of the proper arguments for the isolate creation was moved to the lowest
level possible, where it belongs.

Change-Id: I5b8090fa41290f93a32405d9cfc32dfc76013062