Bug#37489167  Update Rollback on BLOB with INSTANT
Issue:
  On tables having INSTANT ADD columns, rollback of an update introduces
new fields. The update vector pre-allocates the fields array based on
the number of fields in the undo log. Adding new fields to this array
did not account for size exceeding allocated capacity.

Fix:
  When appending a new field to the fields array, check if sufficient
space is already allocated to accommodate the new field. If not, ensure
to allocate larger capacity and copy the old fields to the new fields
array.

Change-Id: I068009bff6419c201817b4917e3a4f85b9f5d3d2