Bug#37523469: 8% regression (around 100k tps loss) seen in sysbench-lua test of point-select during 9.2 release
The WL#16572 implementation introduced an attempted call to the event notification
callbacks for the sha2 plugin and some wrong infrastructure trying to do
this as if it was the component->plugin API bridge for events.

This was not necessary since the sha2 notiifcation callbacks can be
called by the normal reference cache machinery.
But they weren't, because the reference cache was not tracking
notification implementations that were not registered by a component's
provides_service list.

Reverted that part of the worklog with the wrong plumbing.
Added a notification to the reference cache callback for services
manually registered and unregistered through registry_registration
instead.

No visible behavior change, hence no new tests.

Change-Id: I57afcd3c1cc42bef39cd3756a0b2f153cbd5bf5f