Bug#37884336: Assertion failure in Query_block::apply_local_transforms
The problem here was an assert in Query_block::apply_local_transforms.
The function sometimes needs to have the current query block registered
in the LEX, however we had a failing assert for this property.
The reason for this was that function
transform_table_subquery_to_join_with_derived() didn`t restore this
properly after invocation.

The fix is to ensure that current query block is correctly restored
after such call.

Change-Id: I277147832a99e47eddaf86d44a766ff84f133a1f