Bug#37993583: Avoid exclusive MDL during secondary load [noclose]
Issue: EXCLUSIVE MDL during secondary_load is too restrictive and
creates performance issues in the presence of long running queries: the
load needs to wait for the running query to complete, in the meantime
more queries may be admitted at InnoDB and the system can give the false
impression that is hung.

Fix: Apply the fix of Bug#32508608, where instead of an EXCLUSIVE MDL,
we get a SHARED_NO_WRITE lock. This patch was reverted due to some wrong
behavior of SecondaryEngine's Global State. However, there is no strong
evidence that Bug#32508608 was the culprit and the issue cannot be
reproduced in current trunk.

The solution of Bug#32508608 was still taking a XMDL at the end of the
load in order to purge TDC. However, the only thing that secondary_load
modifies in the TABLE_SHARE is the secondary_load flag. The modification
of this flag does not take place during recovery and during self load,
allowing us to skip TDC removal and completely avoid the XMDL for these
cases.

Change-Id: I821fd6389e148ca0401cbce1ef851d5a326aedc6