Bug#37868740  Signal 8 received; Floating point exception
Context:
In SPJ block, after every fragment scan,scanFrag_getBatchSize()
function is used to estimate the number of additional rows that
can be fetched into client batch buffers.
This estimation considers the remaining available 'bytes' and
'rows' configured for the batch, as well as the 12-bit limit
of correlation IDs.
The smallest of these three limits determines the calculated
batch size.
If any of these limits are zero, then the calculated BatchSize
will also be zero, as it will be the minimum of the available
options.
The batch size is then used by scanFrag_parallelism to
calculate the parallelism for the next fragment scan.
If the parallelism cannot be estimated, it returns 0. In
particular, if BatchSize is zero parallelism will also be zero.

Problem:
Division by zero in the calculation of batchSize_rows when
estimated parallelism in zero.

Solution:
Stop requesting fragments for current batch when
scanFrag_getBatchSize returns 0 (No batch buffer left)
or when the estimated parallelism is zero.

Change-Id: I17261df2f95f2bfc1041ca4ea1cf3bad8da24a2f