Bug#36974092 (4/4) Cluster/J has one ClusterConnection per database
Step 4: Move VariableByteBufferPool and remove legacy cleaners

The ownership of a VariableByteBufferPool, which was left in place in
the ClusterConnection in step 1, is now moved from ClusterConnection
to DbFactory. All the code supporting legacy cleaners and finalizers is
removed from VariableByteBufferPool.

Moving the ByteBufferPool out of ClusterConnection allows dropping
the reference to the ClusterConnection from several classes that
kept one, including DictionaryImpl and ResultDataImpl.

The VariableByteBufferPool itself now lives in a static map. This
finally allows two SessionFactories to share an underlying connection
but use differently configured ByteBufferPools. The hash code of the
array spec that specifies the pool configuration is added to the
SessionFactory key. Taking advantage of this in ByteBufferPoolTest
allows the elimination of two disconnect/reconnect cycles that were
previously required to run the test.

Some simple refactoring (and refinement to ClearSmokeTest) allows
removal of another disconnect/reconnect cycle in SessionFactoryTest.

Change-Id: I820c2905dd5c40b2f9ebf31be50418d5708d28fe