Bug#32474777 RonDB compatible session cache
This patch adds a session cache inspired by (but quite different from)
the session cache in the RonDB contribution.

The session cache as implemented here is in essence a cache of Ndb
objects, maintained in DbFactoryImpl.

The instance cache ("DTO cache") from the contributed patch has not
been incorporated into Cluster/J at this time.

Public API changes under clusterj-api include a new SessionFactory
property com.mysql.clusterj.max.cached.sessions. The session cache is
enabled by default, and the default maximum size is 100. Three new
methods in Session are also added as wrappers, for compatibility with
existing releases of RonDB, but marked as deprecated. These changes
are documented in comments under clusterj-api/ and should be added to
the Cluster/J developer documentation.

Change-Id: Ic20de14f0dd55ba8886e7a4d9ca9cdb9ad68eb3e