Bug#37792409 RonDB compatible MultiDB
This patch adds support in SessionFactory for a new version of
getSession() that takes a String database name as a parameter,
so that a single SessionFactory can now handle more than one database.

Because the implementation is built on top of the global connection
pool added for bug#36974092, it is more compact than the one
contributed with RonDB. It relies on an "umbrella" SessionFactory
that manages a set of underlying per-database SessionFactory objects.

The new behavior is not enabled by default; it must be enabled by
setting property com.mysql.clusterj.multi.database to true.

The public API changes under clusterj-api/ should be added to the
Cluster/J developer documentation.

Change-Id: I63f4cf1fbadf76115c1eff754ef6a07d1d121eea