Bug#37861635 Cluster/J schema change handling fails with many threads
This changes provides APIs for improved handling of schema changes
from Cluster/J applications.

This includes 3 new public methods on ClusterJDatastoreException:
 boolean isStaleMetadata()
 boolean isSchemaChangePending()
 void awaitSchemaChange()

When Cluster/J detects that a schema change has caused a
domain type handler to become invalid, it will throw
ClusterJDatastoreException. The user can call isStaleMetadata()
on this exception, and, if this returns true, should then call
session.unloadSchema(). The first thread to enter unloadSchema()
will use the invalid type handler as a lock. If isStaleMetadata()
returns false, the user should call isSchemaChangePending() --
and, then, if this is true, can use awaitSchemaChange() to wait
for the lock to be released.

In SessionFactoryImpl, typeToHandlerMap, formerly a HashMap,
becomes a ConcurrentHashMap. This improves concurrency for the most
common case -- threads that request an existing, cached, domain
type handler -- but requires other synchronization for the less
common case of creating a new handler.

When a table schema changes, Cluster/J will now write a log
message containing the former and current version numbers of
the table metadata. This number is in a dotted form that
consists of both major and minor version numbers.

Change-Id: I02626d628638b403daa189ffd0c2343442b18e03