Bug#37687485 ndb_restore : 626 Tuple did not exist
Problem
ndb_restore receiving unexpected Permanent error 626 when performing
restore of data from DATA files, using NdbApi write operations,
which should semantically never fail with this error.

Root cause

A bug in an error handling path in DBLQH block for handling the case
where :
 - The LQH instance has exhausted the RedoBuffer for
   the Redo log part
 - The DefaultRedoProblemAction is ABORT.
   (Default is QUEUE)
 - The RedoBuffer exhaustion has some component which is
   'booked' space rather than purely being consumed space

In this circumstance, the exhaustion is detected based on the
consumed space + booked space, but the mapping to error codes
is based only on the consumed space.
If the consumed space alone is not sufficient then *no* error
code is set.
DBLQH uses a block member variable to hold the most recently
set error, which is sent back to the requesting block in an
LQHKEYREF signal.  In this case where a problem is encountered,
but no specific error is set, the requesting block is sent
an error code from previous processing.

Solution

Use the logic for deciding that there is Redo log buffer
exhaustion to determine the error code to generate.

Avoid accidentally taking a path intended for error
insertion which can lead to no error being set.

Testing

Tested manually

Change-Id: Ie0badc9c7e9a2084fd7e831a6af521777b368cd2