Bug#37866737 Ndb : TransientPool debugging enhancements
For several classes in data node memory pool implementation a similar
error message "Magic::match failed in ..." can be generated.  The class
name is added in all of them to make it easier to see which instance
raised the error.

The log message also included a i-value, but in for StaticSlotPool the
i-value was decomposed into page number and page index and the page
index was logged instead of whole i-value. For clarity that was changed
such that the whole i-value and both its components were logged.

Change-Id: Ic23981862306767c8a80df87dbc854ee311a0aec